Create `pkg/create`
Use `pkg/create` to create different cluster resource based on user’s input
