+++ cmd/dirigent.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"github.com/spf13/cobra"
+	flag "github.com/spf13/pflag"
 )
 
 func Execute() error {
@@ -19,4 +20,12 @@ func AddCommands() {
 	DirigentCmd.AddCommand(deployCmd)
 	DirigentCmd.AddCommand(versionCmd)
 	DirigentCmd.AddCommand(createCmd)
+}
+
+func flagChanged(flags *flag.FlagSet, key string) bool {
+	flag := flags.Lookup(key)
+	if flag == nil {
+		return false
+	}
+	return flag.Changed
 }
