+++ pkg/deploy/deploy.go
@@ -2,11 +2,7 @@ package deploy
 
 import (
 	log "github.com/Sirupsen/logrus"
-	"github.com/docker/libcompose/docker"
-	"github.com/docker/libcompose/project"
-	"github.com/ghodss/yaml"
-	"io/ioutil"
-	"os"
+	"github.com/weitenghuang/dirigent-cli/pkg/utils"
 	"strings"
 )
 
@@ -21,7 +17,7 @@ const (
 )
 
 func Run(filename string) error {
-	composeObject, err := ParseDockerCompose(filename)
+	composeObject, err := utils.ParseDockerCompose(filename)
 	if err != nil {
 		return err
 	}
@@ -30,12 +26,12 @@ func Run(filename string) error {
 	for _, name := range composeServiceNames {
 		if composeServiceConfig, ok := composeObject.ServiceConfigs.Get(name); ok {
 			log.Infoln("Replication Controller Deployment Starts: ", name)
-			if err := DeployReplicationController(name, composeServiceConfig); err != nil {
+			if err := ReplicationController(name, composeServiceConfig); err != nil {
 				log.Errorln("Error: Deploy ReplicationController ", err, name, composeServiceConfig)
 				return err
 			}
 			log.Infoln("Service Deployment Starts: ", name)
-			if err := DeployService(name, composeServiceConfig); err != nil {
+			if err := Service(name, composeServiceConfig); err != nil {
 				log.Errorln("Error: Deploy Service ", err, name, composeServiceConfig)
 				return err
 			}
@@ -46,44 +42,6 @@ func Run(filename string) error {
 	return nil
 }
 
-func ParseDockerCompose(filename string) (*project.Project, error) {
-	context := &docker.Context{}
-	if filename == "" {
-		filename = "docker-compose.yml"
-	}
-	context.ComposeFiles = []string{filename}
-	composeObject := project.NewProject(&context.Context, nil, nil)
-	err := composeObject.Parse()
-	if err != nil {
-		log.Fatalf("Failed to load compose file", err)
-		return nil, err
-	}
-	log.Infof("Post-parsing: %#v\n", composeObject)
-	return composeObject, nil
-}
-
-func DocodeDockerComposeYaml(filename string) (map[string]interface{}, error) {
-	if _, err := os.Stat(filename); os.IsNotExist(err) {
-		return nil, err
-	}
-
-	dockerComposeYaml, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-
-	var dockerCompose map[string]interface{}
-	if err := yaml.Unmarshal(dockerComposeYaml, &dockerCompose); err != nil {
-		return nil, err
-	}
-	log.Infof("%#v\n", dockerCompose)
-	return dockerCompose, nil
-}
-
 func getPodSelectorLabel(appName string, version string) string {
 	return strings.Join([]string{appName, "-", version, "-pod"}, "")
-}
-
-func getRCSelectorLabel(appName string, version string) string {
-	return strings.Join([]string{appName, "-", version, "-rc"}, "")
 }
